BACKEND-21 Merge `BlockStream` and `TransactionStream` - Removed `BlockStream`, `TransactionStream`, `BlockPublisher`, and `TransactionPublisher` - Added a new Kinesis stream: `TransactionBlockStream` - Added a new Lambda function: `TransactionBlockPublisher` - Added a new log group and role for the Lambda function
