--- stream/src/TransactionPublisher/requirements.txt
@@ -1,1 +1,0 @@
-pymysql
